272 -> 38 speed improvement by changing nBitsSent to be 16 bits instead of 32

+++ TLC5940/load.cpp
@@ -83,7 +83,7 @@ void loadingPrepareUpdate(uint8_t idxHorizontalPixel) {
 		
 		if ( (idxHorizontalPixel<HORZ_PIXELS) && (idxHorizontalPixel<GRAPHIC_WIDTH)) {
 
-			uint32_t nBitsSent = 0;
+			uint16_t nBitsSent = 0;
